tests/wpt/tests/html/semantics/forms/the-input-element/defaultValue-clobbering.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" title="Joey Arhar" href="mailto:jarhar@chromium.org">
<meta name="assert" content="Assigning to defaultValue does not modify text a user has already typed in.">

<!-- This behavior is not explicitly specified. -->

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<div>
  email with leading whitespace: <input id=emailinput type=email>
</div>
<div>
  number with trailing incomplete exponent: <input id=numberinput type=number>
</div>

<script>
promise_test(async () => {
  await test_driver.send_keys(emailinput, '  user');
  assert_false(emailinput.validity.valid, '"  user" should not be a valid value for type=email.');

  emailinput.defaultValue = emailinput.value;
  assert_false(emailinput.validity.valid, 'Assigning to defaultValue should not affect input.validity.');
}, 'Visible value and validity should not be affected when assigning to the defaultValue property for type=email.');

promise_test(async () => {
  await test_driver.send_keys(numberinput, '123e');
  assert_false(numberinput.validity.valid, '"123e" should not be a valid value for type=number.');

  numberinput.defaultValue = numberinput.value;
  assert_false(numberinput.validity.valid, 'Assigning to defaultValue should not affect input.validity.');
}, 'Visible value and validity should not be affected when assigning to the defaultValue property for type=number.');
</script>
